Write workflow prototype.

+++ core/src/main/java/tachyon/worker/block/allocator/Allocator.java
@@ -6,5 +6,5 @@ import tachyon.worker.block.meta.BlockMeta;
  * Interface for the allocation policy of Tachyon managed data.
  */
 public interface Allocator {
-  BlockMeta allocateBlock(long userId, long blockId, long blockSize, int tierHint);
+  BlockMeta allocateBlock(int userId, long blockId, long blockSize, int tierHint);
 }
